# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Basic API",
        "description": "L5 Swagger API with Laravel basic",
        "contact": {
            "email": "nguyen.an.thoi.212@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/laravel_basic/an.thoi.laravel/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Profile User",
                "description": "Get Profile User",
                "operationId": "GetProfileUser",
                "responses": {
                    "200": {
                        "description": "Get profile Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/edit-profile": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit Profile User",
                "description": "Edit Profile User",
                "operationId": "putEditProfile",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Edit profile user",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "default": "admin 123"
                                },
                                "gender": {
                                    "type": "boolean",
                                    "default": false
                                },
                                "dateOfBirth": {
                                    "type": "string",
                                    "default": "10/10/2010"
                                },
                                "about": {
                                    "type": "string",
                                    "default": "T la admin"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edit profile successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "A short description for security scheme",
            "name": "Authorization",
            "in": "header"
        }
    }
}